requirements.txt
youtube_transcript_api
streamlit
groq
python-dotenv
pathlib
requests
